修复数据同步bug：数据库到数据库跨库同步无效
